Split Point to its own file
