+++ include/NAS2D/Resources/FontInfo.h
@@ -9,7 +9,7 @@
 // ==================================================================================
 #pragma once
 
-#include "../Renderer/Primitives.h"
+#include "../Renderer/Point.h"
 
 #include <iostream>
